+++ CHANGELOG.md
@@ -1,11 +1,24 @@
 # Changelog
 
-## [Unreleased
-**Date:** 2021-07-27
+## Release 0.12.4
+**Date:** Unreleased
 
+* Starlit is now used for notebook testing in starboard-notebook repo.
+* ESBuild loader plugin for webpack is now used (cutting build time in half or so).
+* Added `build:nominify` option for faster building (and easier debugging).
 * Add [COOP/COEP headers](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/SharedArrayBuffer#security_requirements) to enable the use of the `SharedArrayBuffer` and `Atomics` APIs.
 * Add optional COOP/COEP enabling service worker. Useful when you cannot change the HTTP headers that the server sends.
 
+## Release 0.12.3
+**Date:** 2021-07-23
+
+* The initial cell run (when the notebook is loaded) now sets `isInitialRun: true` to true on the event `sb:run_all_cells` event.
+
+## Release 0.12.2
+**Date:** 2021-07-23
+
+* Added `sb:run_all_cells` which gets fired when `runtime.controls.runAllCells` is called. This allows you to cancel that action or otherwise act on it.
+* Fixed `runtime.exports.libraries.litDirectives` not actually containing the `lit` library directives in the bundle.
 
 ## Release 0.12.1
 **Date:** 2021-06-18